[2단계 - 로또(수동)] 알파(배철환) 미션 제출합니다. (#420)
* refactor : 테스트 코드에 프로덕션 코드의 패키지 구조를 도입

* docs : 추가된 기능목록 추가

* feat : 수동 로또 구매 갯수 입력 기능 추가

* feat : 수동 로또 갯수가 전체 로또 갯수보다 작거나 같은지 검증하는 기능 추가

* feat : 로또 한 장을 전체 로또티켓에 추가하는 기능 추가

* feat : 수동 로또 번호 입력 기능 추가

* docs : 구현된 기능 업데이트

* feat : Lotto 클래스의 단위 테스트 추가

* fix : 총 당첨금액의 오버플로우 문제로 long으로 변경

* refactor : 수동입력 기능을 분리

* refactor : 수동입력갯수가 전체 로또 갯수보다 많은 지 검증 시 파라미터 제거

* feat : 로또 구매 갯수와 수동 갯수를 담당하는 도메인 추가

+++ src/main/java/domain/StatisticCalculator.java
@@ -24,12 +24,12 @@ public class StatisticCalculator {
 	}
 
 	public double calculateProfitRatio() {
-		int totalPrice =
+		long totalPrice =
 			this.resultStatistics
 				.keySet()
 				.stream()
 				.filter(num -> this.resultStatistics.get(num) > 0)
-				.mapToInt(this.resultStatistics::get)
+				.mapToLong(this.resultStatistics::get)
 				.sum() * UNIT_PRICE;
 		int totalPrize = this.resultStatistics.keySet().stream().filter(num -> this.resultStatistics.get(num) > 0)
 			.mapToInt(num -> num.getProfit() * this.resultStatistics.get(num)).sum();
